feat(workflow): interrupt in-flight turns at the workflow budget ceiling (GH-1770)
Fourth enforcement increment from specs/GH1770/tech.md §4.3, after the
pre-dispatch gate (#1893), the daily profile cap (#1903), and the
completion ceiling (#1904). Dispatch-time gating cannot catch an
activity that blows its budget mid-turn, and the completion ceiling
only reacts once the turn already finished spending.

- RuntimeUsageContext carries the store's RuntimeBudgetPolicy and, right
  after each streamed TokenUsage is persisted, re-reads the workflow
  spend so the aggregate includes the usage that may have crossed the
  ceiling. Comparison stays in integer micro-dollars.
- enforce returns a TurnBudgetStop from process_stream_item; the turn
  loop interrupts the agent through the same adapter.interrupt() path
  the lease-lost branch uses and ends the turn with a budget message.
- shadow records a BudgetShadowDecision event
  (source workflow_runtime_turn_watchdog, decision would_interrupt) and
  lets the turn run.
- A failed ceiling read is logged at error level rather than silently
  disabling the watchdog; the completion ceiling remains the backstop.
- turn_lifecycle.rs was at 893 lines, over the repository ceiling, so
  its inline test module moved to turn_lifecycle_tests.rs, matching the
  sibling stall_tests.rs / terminal_error_tests.rs convention. No test
  bodies changed.
- Postgres-gated tests: enforce stops at the ceiling, shadow records the
  event without stopping, under-budget turns keep running, and the
  unlimited opt-out disables the watchdog.
